# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Full Sales Web System API
  version: 1.0.0
  description: Driver/seller control system — full REST surface (see docs/API-CONTRACT.md)
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://127.0.0.1:8080/v1
    description: Local development
paths:
...
  /sales/{id}:
    get:
      summary: Get sale by id
      operationId: getSale
      tags: [Sales]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SaleId'
      responses:
        '200':
          description: Sale found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/SaleNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SaleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    SaleNotFound:
      description: Sale not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorCode:
      type: string
      description: Machine-readable error codes returned in error.code
      enum:
        - NOT_FOUND
        - UNAUTHORIZED
        - INVALID_CREDENTIALS
        - FORBIDDEN
        - INVALID_CNPJ
        - VALIDATION_ERROR
        - SALE_NOT_FOUND
        - INSUFFICIENT_STOCK
        - INVALID_SALE_TRANSITION
        - COMMERCE_NOT_FOUND
        - PRODUCT_NOT_FOUND
        - COMMERCE_INACTIVE
        - INACTIVE_PRODUCT
        - ORDER_NOT_FOUND
        - INVALID_ORDER_TRANSITION
        - INVALID_DELIVERY_ADDRESS
        - REJECTION_REASON_REQUIRED
        - USER_NOT_FOUND
        - UNAUTHORIZED_PAYMENT_DECLARATION
        - INSUFFICIENT_BALANCE
        - MEDIA_NOT_FOUND
        - INVALID_MIME
        - FILE_TOO_LARGE
        - DELIVERY_NOT_FOUND
        - PROOF_REQUIRED
        - INVALID_DELIVERY_TRANSITION
        - REPORT_NOT_FOUND
        - SIGNING_KEY_UNAVAILABLE
        - INTERNAL_ERROR
        - REASON_REQUIRED
        - INVALID_MOVEMENT_TYPE
        - DELIVERY_EXISTS
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          additionalProperties: false
          properties:
            code:
              oneOf:
                - $ref: '#/components/schemas/ErrorCode'
                - type: string
              example: SALE_NOT_FOUND
            message:
              type: string
              example: Sale not found
            correlationId:
              type: string
              format: uuid
...
    PaymentMethodResponse:
      type: string
      enum: [Cash, Pix, Credit, Debit]
    SaleStatus:
      type: string
      enum: [Pending, Confirmed, Cancelled]
...
    Sale:
      type: object
      required:
        - id
        - commerceId
        - driverId
        - status
        - paymentMethod
        - totalAmount
        - totalCurrency
        - items
      properties:
        id:
          type: string
          format: uuid
        commerceId:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SaleStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodResponse'
        totalAmount:
          type: integer
          description: Total in minor units (centavos)
        totalCurrency:
          type: string
          example: BRL
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
    SaleItem:
      type: object
      required:
        - productId
        - quantity
        - unitPriceAmount
        - unitPriceCurrency
        - lineTotalAmount
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
        unitPriceAmount:
          type: integer
        unitPriceCurrency:
          type: string
        lineTotalAmount:
          type: integer
...